requirements.txt
alembic==1.7.7
click==8.1.2
environs==9.5.0
Flask==2.1.1
Flask-HTTPAuth==4.5.0
Flask-JWT-Extended==4.3.1
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
importlib-metadata==4.11.3
itsdangerous==2.1.2
Jinja2==3.1.1
Mako==1.2.0
MarkupSafe==2.1.1
marshmallow==3.15.0
packaging==21.3
psycopg2-binary==2.9.3
PyJWT==2.3.0
pyparsing==3.0.8
python-dotenv==0.20.0
SQLAlchemy==1.4.35
Werkzeug==2.1.1
zipp==3.8.0
